requirements.txt
# FastAPI and core dependencies
fastapi[standard]==0.117.1
uvicorn[standard]==0.31.0
pydantic==2.9.2
pydantic-settings==2.6.0

# Database dependencies
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3

# Testing dependencies
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0
factory-boy==3.3.0
faker==20.1.0
respx==0.20.2

# Development dependencies
black==24.8.0
isort==5.13.2
flake8==7.1.1
pre-commit==3.8.0

# Additional utilities
python-multipart>=0.0.18
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Monitoring and metrics
psutil==6.0.0